feat: working GET (simple blocking client)
